Add failed program detection. Bump to version name 2015. Fix text clock parse int exception issue. Sync service now cleans program content e.g.,new.files/ only for USB synced. Fix crash issue when detecting the video tracks by commenting out the detection on program parsing. For the loop feature of some video, as the track detection is remove, using video file extension. Currently only ".mov" and ".mp4" are valid for video looping.

+++ color/src/main/java/com/color/home/widgets/clocks/ItemTextClock.java
@@ -625,7 +625,9 @@ public class ItemTextClock extends TextView {
         setBackgroundColor(GraphUtils.parseColor(item.backcolor));
 
         // <ftColor>4294967295</ftColor>
-        setTextColor(Integer.parseInt(digitalClock.ftColor) | 0xFF000000);
+        setTextColor((int)(Long.parseLong(digitalClock.ftColor)) | 0xFF000000);
+        // It's not 0xFFFFFFFF, so don't use the following.
+        // setTextColor(GraphUtils.parseColor(digitalClock.ftColor) | 0xFF000000);
 
         // We take the 0xFF000000 (BLACK) to transparent.
         if (!TextUtils.isEmpty(item.backcolor) && !"0xFF000000".equals(item.backcolor))
